Use market://-links when page is opened on an Android device
By using market://-links instead of Google Play links on Android devices
the user is able to install the application from any app store available
on his device. This makes sense since the Google Play store is not the
only place where KISS may be installed from.

This change is implemented by adding a piece of JavaScript that runs just
after the page's content has been downloaded and, if the User-Agent
indicates that this is an Android browser, replaces all Google Play links
with their corresponding market://-links